Factor out fractional part of float
As we need to communicate whether thread was not float (i.e. we are
handling an integer number) or the fractional part was invalid (e.g. a
dot with no digits following), a simple boolean for a match did not
suffice anymore. Hence, Match_Type was introduced which can signal a
match, no match or invalid data.